🛑 WIP support for expandable structs (see #21, #19)
some structs carried by messages or attributes are being expanded in
newer kernel versions. introduce metadata
to tag such structs in the type definitions, along with the fields
present in different ABI versions. use it on the problematic structs.

then, update the code generation to produce more correct code for these
structs. right now it's very limited: it only
parses the 1st version of the struct and adds trailing
data as an `__unparsed` field. this data is preserved when reformatting
the struct.

in the future it should interfaces for each ABI version, and select
(depending on the length) which one to use for parsing. it's not clear
though what to do for formatting...

in any case, a breaking side effect of this change is that the
`rxNoHandler` field (added in kernel 4.6) is
no longer parsed. in return, this library now works on <4.6 and >5.18.